+++ src/internal/common/currencyconversion/currencyconversion.go
@@ -3,5 +3,5 @@ package currencyconversion
 type CurrencyConversion struct {
 	FromCurrency   string  `json:"from_currency"`
 	ToCurrency     string  `json:"to_currency"`
-	ConversionRate float32 `json:"conversion_rate"`
+	ConversionRate float64 `json:"conversion_rate"`
 }
