+++ src/internal/common/fetcherresponse/fetcherresponse.go
@@ -3,5 +3,5 @@ package fetcherresponse
 type FetcherResponse struct {
 	Date  string  `json:"date"`
 	Quote string  `json:"quote"`
-	Rate  float32 `json:"rate"`
+	Rate  float64 `json:"rate"`
 }
