+++ src/internal/common/queryresult/queryresult.go
@@ -7,21 +7,21 @@ type Query1Result struct {
 	FromAccount string  `json:"from_account"`
 	ToBank      string  `json:"to_bank"`
 	ToAccount   string  `json:"to_account"`
-	Amount      float32 `json:"amount"`
+	Amount      float64 `json:"amount"`
 }
 
 type Query2Result struct {
 	BankName    string  `json:"bank_name"`
 	FromBank    string  `json:"from_bank"`
 	FromAccount string  `json:"from_account"`
-	Amount      float32 `json:"amount"`
+	Amount      float64 `json:"amount"`
 }
 
 type Query3Result struct {
-	FromBank    string  `json:"from_bank"`
-	FromAccount string  `json:"from_account"`
+	FromBank      string  `json:"from_bank"`
+	FromAccount   string  `json:"from_account"`
 	PaymentFormat string  `json:"payment_format"` // revisar
-	Amount      float32 `json:"amount"`
+	Amount        float64 `json:"amount"`
 }
 
 type Query4Result struct {
@@ -57,7 +57,7 @@ func (q Query1Result) GetHeaders() []string {
 }
 
 func (q Query2Result) GetHeaders() []string {
-	return []string{"From Bank", "Account", "Bank Name" , "Amount Paid"}
+	return []string{"From Bank", "Account", "Bank Name", "Amount Paid"}
 }
 
 func (q Query3Result) GetHeaders() []string {
